requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
Authlib==1.3.1
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
groq==0.18.0
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.34
langchain-groq==0.2.4
langsmith==0.3.8
orjson==3.10.15
packaging==24.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
service-identity==24.2.0
setuptools==75.8.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.0.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validators==0.34.0
weaviate-client==4.10.4
zope.interface==7.2
zstandard==0.23.0
